auth-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
email-validator
PyJWT
passlib[bcrypt]
